03-advanced-techniques/03-context-engineering/pyproject.toml
[project]
name = "lesson-context-engineering"
version = "0.1.0"
description = "Context window management with token counting and compression"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.76.0",
    "python-dotenv>=1.0.0",
    "common",
    "rich>=13.7.0",
]

[tool.uv.sources]
common = { path = "../../common", editable = true }

[tool.ruff]
line-length = 100
target-version = "py311"
